22/10/2024 dashboard completed; buy basket; delete basket;

+++ deleteBasket.php
@@ -1,8 +1,10 @@
 <?php
-
 include 'Connections\config.php';
-$q = mysqli_query($conn, "SELECT FROM basket WHERE status = 2");
-$a = mysqli_fetch_array($q);
-$basket_id = $a["id"];
-mysqli_query($conn, "DELETE FROM basket WHERE id = '$basket_id'");
-echo '<meta http-equiv="refresh" content="0;url=index.php?nav=productDisplay">';
+if (isset($_REQUEST["basket_id"])) {
+    $user_id = $_REQUEST["user_id"];
+    $basket_id = $_REQUEST["basket_id"];
+    mysqli_query($conn, "DELETE FROM basket WHERE id = '$basket_id'");
+    mysqli_query($conn, "DELETE FROM product_basket WHERE basket_id = '$basket_id'");
+    echo '<meta http-equiv="refresh" content="0;url=index.php?nav=productDisplay">';
+}
+
